Add support for OPTIONS HTTP requests
This change modifies the Nginx site configuration so that OPTIONS
requests pass through to index.php (ie. Flow or Neos), in order to
enable Flow to react on such requests.

Notice that directory listing will not work anymore this way, because
we remove "…/" from the try_files clauses. In practice this should
not be a problem though, because we disallow display directory indexes
anyway.