Add query parameter for GCS cache busting
Google may store storage objects in various edge caches according to
the given cache-lifetime controlled via Cache-Control and Expires.
Therefore, if metadata was changed, these changes may not be visible
to this Nginx. In order to get always the freshest version, we append
the request id as a query parameter as a "cache buster".